fix(runtime): resolve documentation-review findings VDOC-042..128
Continues the systematic verification of docs/documentation-review-findings.md.
Every finding in the range was checked against the current implementation,
fixed (or confirmed already fixed), stamped in the findings document with a
resolution note, and covered by a regression test where practical.

Text/pattern engine (VDOC-042..064):
- Regex: NUL-containing patterns trap instead of silently truncating and
  aliasing the compile cache; zero-width Replace/Split use ECMAScript-style
  semantics (no dropped source bytes); complement shorthands union their byte
  sets so mixed classes like [a\D] match correctly; concat-level backtracking
  sees through groups so (a*)a matches like a*a; the capture matcher is now a
  CPS engine with lexical group numbering, restore-on-backtrack slots, and no
  32-group cap (dynamic arrays sized from the compiled pattern).
- FuzzyMatch clamps successful scores at >= 0 so -1 reliably means a miss.
- String.Like performs a strict UTF-8 decode (overlongs, surrogates, and
  out-of-range code points trap).
- Diff.Patch validates the diff against its original argument and traps on
  mismatch; Markdown.ToText preserves intraword underscores; markdown line
  advancement no longer forms out-of-bounds pointers.
- Serialize/TextWrapper/Template stale comments corrected; case-insensitive
  helpers now use a locale-independent ASCII layer (new rt_ascii.h) so results
  cannot change with the host LC_CTYPE; platform-policy lint is strict-clean.

Localization (VDOC-065..085):
- Locale parser implements the RFC 5646 well-formedness grammar (private-use
  tags, extlang, empty-extension/ordering/duplicate-variant rejection, 128-byte
  canonical buffer); FromParts requires one pre-split subtag per argument;
  Equals treats null as invariant.
- Locale JSON loading rejects embedded NUL escapes, validates digit strings
  with a strict UTF-8 decoder, and verifies formatter data at load time
  (date-pattern letters, exactly-one currency placeholders, distinct
  separators, {n} in relative-time templates, {0}/{1} in list templates,
  collation strength 1..3).
- NumberFormat: lenient parsing rejects empty grouping runs; Scientific uses
  locale sign/NaN/infinity tokens; formatted non-finite values round-trip
  through the parsers; non-default CurrencyOf output parses back.
- RelativeTime keeps INT64_MIN's exact magnitude; plural rules evaluate
  integer operands in exact 64-bit space and derive correct operands from
  exponent-notation renderings; ListFormat and PluralRules.Categories no
  longer leak references; MessageBundle.Plural uses the locale digit set and
  full-chain fallback cycle detection; the C-locale cache init is race-free.

Collections (VDOC-086..107):
- Queue/Stack/Deque/Ring membership uses boxed-value equality; Queue/Stack
  expose SetOwnsElements; default sorting is one shared total/transitive
  order (type-class rank, then value); Bag.ToSet boxes entries so value
  lookups match.
- LazySeq: typed receivers (new class id + checked casts), owning collectors,
  retained Repeat values, consistent bound validation, alloc-failure trap, and
  Option-returning Next/Peek; iterators are registered with the cycle
  collector; Trie gains LongestPrefixOption.
- BitSet staged growth keeps spare capacity invisible (Count <= Length holds;
  binary ops stay in bounds); I64Buffer arithmetic traps on overflow;
  CountMap.IncBy honors its returned-count contract; Bytes constructor
  metadata agrees across the manifest; missing count-ceiling guards added.

Zia tooling (VDOC-108..116):
- Weak-stub IsDone matches the strong service; document mirror deltas enforce
  revision ordering and strict range validity with bounded numeric parsing;
  document symbols cover only the active file; legacy completion rows escape
  multiline snippets; stub payloads carry the full structured schema;
  IsAvailable and Document.Has disambiguate absence from clean results;
  rename validation mirrors the lexer exactly; async diagnostics carry the
  complete fixits sequence.

Audio (VDOC-117..124):
- Audio-disabled mix-group state is lock-protected; over-long group names are
  rejected instead of aliasing; playlists skip unloadable entries with a
  bounded scan; spatial playback uses the same -1 failure sentinel as Sound;
  SoundBank registration requires playable wrappers; peaking EQ sanitizes its
  gain; Sound.LoadAsset returns null for missing assets; SpatialAudio3D
  signatures carry typed Sound/Vec3 parameters.

Threads (VDOC-125..128):
- VM Pool.Submit honors shutdown state on both interpreted backends;
  Parallel ForEach/Map/Reduce now run (sequentially) on both VMs instead of
  trapping; Async.RunOwned has real VM bridges with native-parity ownership
  (Run borrows everywhere) and the remaining callback variants trap safely
  instead of casting VM function values as native code; Pool.SubmitOwned adds
  an owned-argument submission whose argument is released on run or discard.

Also fixes a latent test expectation (Template.Keys is StringSet-typed) and
registers all new symbols with the runtime surface policy and baselines.

+++ src/runtime/collections/rt_bloomfilter.h
@@ -49,7 +49,7 @@ void rt_bloomfilter_add(void *filter, rt_string item);
 /// @param filter Bloom filter object.
 /// @param item String to check.
 /// @return 1 if possibly present, 0 if definitely not present.
-int64_t rt_bloomfilter_might_contain(void *filter, rt_string item);
+int8_t rt_bloomfilter_might_contain(void *filter, rt_string item);
 
 /// @brief Get the number of items added.
 /// @param filter Bloom filter object.
@@ -67,7 +67,9 @@ void rt_bloomfilter_clear(void *filter);
 
 /// @brief Merge two filters (union).
 /// @param filter First filter (modified in place).
-/// @param other Second filter (must have same parameters).
+/// @param other Second filter (must have the same derived bit and hash
+///              counts; constructor arguments that derive to the same shape
+///              are compatible).
 /// @return 1 on success, 0 if filters are incompatible.
 int64_t rt_bloomfilter_merge(void *filter, void *other);
